server/requirements.txt
replicate>=0.24.0
python-dotenv>=1.0.0
web3>=7.0.0
pgpy>=0.6.0
eciespy>=0.4.5
eth-account>=0.13.0
eth-keys>=0.4.0
coincurve>=21.0.0
requests>=2.31.0 